+++ hyp3_gamma/insar/ifm_sentinel.py
@@ -182,6 +182,9 @@ def move_output_files(output, reference, prod_dir, long_output, include_los_disp
         inName = "{}.inc.tif".format(output)
         outName = "{}_inc_map.tif".format(os.path.join(prod_dir, long_output))
         shutil.copy(inName, outName)
+        inName = "{}.inc_ell.tif".format(output)
+        outName = "{}_inc_map_ell.tif".format(os.path.join(prod_dir, long_output))
+        shutil.copy(inName, outName)
 
     if include_look_vectors:
         inName = "{}.lv_theta.tif".format(output)
@@ -292,9 +295,10 @@ def insar_sentinel_gamma(reference_file, secondary_file, rlooks=20, alooks=4, in
 
     wrk = os.getcwd()
     reference_date = reference_file[17:32]
-    reference_date_short = reference_file[17:25]
+    reference = reference_file[17:25]
     secondary_date = secondary_file[17:32]
-    secondary_date_short = secondary_file[17:25]
+    secondary = secondary_file[17:25]
+
     igramName = "{}_{}".format(reference_date, secondary_date)
     if "IW_SLC__" not in reference_file:
         raise GranuleError(f'Reference file {reference_file} is not of type IW_SLC!')
@@ -304,7 +308,7 @@ def insar_sentinel_gamma(reference_file, secondary_file, rlooks=20, alooks=4, in
     pol = get_copol(reference_file)
     log.info("Processing the {} polarization".format(pol))
 
-    #  Ingest the data files into gamma format
+    # Ingest the data files into gamma format
     log.info("Starting par_S1_SLC")
     orbit_files = []
     for granule in (reference_file, secondary_file):
@@ -312,7 +316,7 @@ def insar_sentinel_gamma(reference_file, secondary_file, rlooks=20, alooks=4, in
         par_s1_slc_single(granule, pol, os.path.abspath(orbit_file))
         orbit_files.append(orbit_file)
 
-    #  Fetch the DEM file
+    # Fetch the DEM file
     log.info("Getting a DEM file")
     dem_source = 'GLO-30'
     dem_pixel_size = int(alooks) * 40  # typically 160 or 80; IFG pixel size will be half the DEM pixel size (80 or 40)
@@ -321,15 +325,12 @@ def insar_sentinel_gamma(reference_file, secondary_file, rlooks=20, alooks=4, in
 
     # Figure out which bursts overlap between the two swaths
     burst_tab1, burst_tab2 = get_burst_overlaps(reference_file, secondary_file)
-
     log.info("Finished calculating overlap - in directory {}".format(os.getcwd()))
-    shutil.move(burst_tab1, f'{reference_date_short}/{burst_tab1}')
-    shutil.move(burst_tab2, f'{secondary_date_short}/{burst_tab2}')
+    shutil.move(burst_tab1, f'{reference}/{burst_tab1}')
+    shutil.move(burst_tab2, f'{secondary}/{burst_tab2}')
+
     # Mosaic the swaths together and copy SLCs over
     log.info("Starting SLC_copy_S1_fullSW.py")
-    reference = reference_date_short
-    secondary = secondary_date_short
-
     os.chdir(reference)
     SLC_copy_S1_fullSW(wrk, reference, "SLC_TAB", burst_tab1, mode=1, dem="big", dempath=wrk, raml=rlooks, azml=alooks)
     os.chdir("..")
@@ -359,7 +360,7 @@ def insar_sentinel_gamma(reference_file, secondary_file, rlooks=20, alooks=4, in
     else:
         log.info("Found azimuth offset of {}!".format(offset))
 
-    output = reference_date_short + "_" + secondary_date_short
+    output = f"{reference}_{secondary}"
 
     log.info("Starting s1_coreg_overlap")
     execute(f"S1_coreg_overlap SLC1_tab SLC2R_tab {output} {output}.off.it {output}.off.it.corrected",
@@ -372,7 +373,7 @@ def insar_sentinel_gamma(reference_file, secondary_file, rlooks=20, alooks=4, in
     log.info("Starting phase unwrapping and geocoding")
     unwrapping_geocoding(reference, secondary, step="man", rlooks=rlooks, alooks=alooks)
 
-    #  Generate metadata
+    # Generate metadata
     log.info("Collecting metadata and output files")
 
     os.chdir(wrk)
@@ -418,7 +419,7 @@ def main():
     parser.add_argument("-r", "--rlooks", default=20, help="Number of range looks (def=20)")
     parser.add_argument("-a", "--alooks", default=4, help="Number of azimuth looks (def=4)")
     parser.add_argument("-d", action="store_true", help="Add DEM file to product bundle")
-    parser.add_argument("-i", action="store_true", help="Create incidence angle map")
+    parser.add_argument("-i", action="store_true", help="Create local and ellipsoidal incidence angle maps")
     parser.add_argument("-l", action="store_true", help="Create look vector theta and phi files")
     parser.add_argument("-s", action="store_true", help="Create line of sight displacement file")
     parser.add_argument("-w", action="store_true", help="Create wrapped phase file")